Added a first batch of VDS tests.
I believe this should be its own bundle of tests rather than being
scattered all over.
