+++ UnitTests/H5DTest/H5DTest.cs
@@ -121,12 +121,6 @@ namespace UnitTests
 
             Assert.IsTrue(H5S.close(dspace) >= 0);
             Assert.IsTrue(H5T.close(dtype) >= 0);
-
-#if HDF5_VER1_10
-
-            createVDS();
-
-#endif
         }
 
         [TestInitialize()]
@@ -168,12 +162,6 @@ namespace UnitTests
 
             File.Delete(m_v0_class_file_name);
             File.Delete(m_v2_class_file_name);
-
-#if HDF5_VER1_10
-
-            cleanupVDS();
-
-#endif
         }
 
         private static hid_t m_v0_class_file = -1;
@@ -206,26 +194,6 @@ namespace UnitTests
 
         private static hid_t m_v2_utf8_dset = -1;
 
-#if HDF5_VER1_10
-
-        private static hid_t m_vds_class_file = 1;
-
-        private static string m_vds_class_file_name;
-
-        private static hid_t m_a_class_file = 1;
-
-        private static string m_a_class_file_name;
-
-        private static hid_t m_b_class_file = 1;
-
-        private static string m_b_class_file_name;
-
-        private static hid_t m_c_class_file = 1;
-
-        private static string m_c_class_file_name;
-
-#endif
-
         // Callback for H5D.iterate
         // op_data is a pointer to a counter and we keep adding the elements
         private herr_t DelegateMethod
@@ -241,110 +209,5 @@ namespace UnitTests
             Marshal.WriteInt32(op_data, count);
             return 0;
         }
-
-#if HDF5_VER1_10
-
-        private static void createVDS()
-        {
-            // create files
-            m_a_class_file = Utilities.H5TempFile(ref m_a_class_file_name,
-                H5F.libver_t.LATEST, true);
-            Assert.IsTrue(m_a_class_file >= 0);
-            m_b_class_file = Utilities.H5TempFile(ref m_b_class_file_name,
-                H5F.libver_t.LATEST, true);
-            Assert.IsTrue(m_b_class_file >= 0);
-            m_c_class_file = Utilities.H5TempFile(ref m_c_class_file_name,
-                H5F.libver_t.LATEST, true);
-            Assert.IsTrue(m_c_class_file >= 0);
-            m_vds_class_file = Utilities.H5TempFile(ref m_vds_class_file_name);
-            Assert.IsTrue(m_vds_class_file >= 0);
-
-            //
-            // create target datasets
-            //
-            hid_t dcpl = H5P.create(H5P.DATASET_CREATE);
-            Assert.IsTrue(dcpl >= 0);
-            int fill_value = 1;
-            GCHandle hnd = GCHandle.Alloc(fill_value, GCHandleType.Pinned);
-            Assert.IsTrue(H5P.set_fill_value(dcpl, H5T.NATIVE_INT,
-                hnd.AddrOfPinnedObject()) >= 0);
-
-            hsize_t[] dims = { 6 };
-            hid_t src_dsp = H5S.create_simple(1, dims, null);
-            
-            // A
-            fill_value = 1;
-            hid_t a = H5D.create(m_a_class_file, "A", H5T.STD_I32LE, src_dsp);
-            Assert.IsTrue(a >= 0);
-            Assert.IsTrue(H5D.close(a) >= 0);
-            // B
-            fill_value = 2;
-            hid_t b = H5D.create(m_b_class_file, "B", H5T.STD_I32LE, src_dsp);
-            Assert.IsTrue(b >= 0);
-            Assert.IsTrue(H5D.close(b) >= 0);
-            // B
-            fill_value = 3;
-            hid_t c = H5D.create(m_c_class_file, "C", H5T.STD_I32LE, src_dsp);
-            Assert.IsTrue(c >= 0);
-            Assert.IsTrue(H5D.close(c) >= 0);
-            
-            //
-            // create the VDS
-            //
-            fill_value = -1;
-            hsize_t[] vds_dims = { 4, 6 };
-            hid_t vds_dsp = H5S.create_simple(2, vds_dims, null);
-
-            hsize_t[] start = { 0, 0 };
-            hsize_t[] count = { 1, 1 };
-            hsize_t[] block = { 1, 6 };
-
-            start[0] = 0;
-            Assert.IsTrue(H5S.select_hyperslab(vds_dsp, H5S.seloper_t.SET,
-                start, null, count, block) >= 0);
-            Assert.IsTrue(H5P.set_virtual(dcpl, vds_dsp,
-                m_a_class_file_name, "A", src_dsp) >= 0);
-
-            start[0] = 1;
-            Assert.IsTrue(H5S.select_hyperslab(vds_dsp, H5S.seloper_t.SET,
-                start, null, count, block) >= 0);
-            Assert.IsTrue(H5P.set_virtual(dcpl, vds_dsp,
-                m_b_class_file_name, "B", src_dsp) >= 0);
-
-            start[0] = 2;
-            Assert.IsTrue(H5S.select_hyperslab(vds_dsp, H5S.seloper_t.SET,
-                start, null, count, block) >= 0);
-            Assert.IsTrue(H5P.set_virtual(dcpl, vds_dsp,
-                m_c_class_file_name, "C", src_dsp) >= 0);
-
-            hid_t vds = H5D.create(m_vds_class_file, "VDS", H5T.STD_I32LE,
-                vds_dsp, H5P.DEFAULT, dcpl, H5P.DEFAULT);
-            Assert.IsTrue(vds >= 0);
-            Assert.IsTrue(H5D.close(vds) >= 0);
-
-            Assert.IsTrue(H5S.close(vds_dsp) >= 0);
-            Assert.IsTrue(H5S.close(src_dsp) >= 0);
-            Assert.IsTrue(H5P.close(dcpl) >= 0);
-
-            hnd.Free();
-
-            // close the satellite files
-            Assert.IsTrue(H5F.close(m_a_class_file) >= 0);
-            Assert.IsTrue(H5F.close(m_b_class_file) >= 0);
-            Assert.IsTrue(H5F.close(m_c_class_file) >= 0);
-        }
-
-        private static void cleanupVDS()
-        {
-            Assert.IsTrue(H5F.close(m_vds_class_file) >= 0);
-
-            File.Delete(m_vds_class_file_name);
-            File.Delete(m_a_class_file_name);
-            File.Delete(m_b_class_file_name);
-            File.Delete(m_c_class_file_name);
-        }
-
-#endif
-
     }
 }
