+++ UnitTests/H5VDSTest/H5VDSTest.playlist
@@ -1,0 +1,1 @@
+<Playlist Version="1.0"><Add Test="UnitTests.H5VDSTest.H5Pget_layoutTestVDS1" /><Add Test="UnitTests.H5VDSTest.H5Pget_virtual_countTestVDS1" /><Add Test="UnitTests.H5VDSTest.H5Pget_virtual_vspaceTestVDS1" /><Add Test="UnitTests.H5VDSTest.H5Pget_virtual_srcspaceTestVDS1" /><Add Test="UnitTests.H5VDSTest.H5Pget_virtual_dsetnameTestVDS1" /></Playlist>
